+++ src/plays/tic-tac-toe-game/styles/tic.css
@@ -1,10 +1,3 @@
-
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
 .squareButton{
     width: 5rem;
     height: 5rem;
@@ -21,9 +14,6 @@
 }
 .row{
     display: flex;
-}
-.bg-yellow-500{
-    background-color: #F59E0B;
 }
 .jumpButton{
     padding: 0.5rem 1rem;
@@ -45,12 +35,15 @@
     margin-top: 1rem;
     font-size: 24px;
 }
-header{
+.header{
     width: 100%;
     padding: 16px 8px;
     color: white;
     font-size: 24px;
-    text-align: center;
+    
+    display: grid;
+    place-content: center;
+
     background-color: #010326;
     margin-bottom: 2rem;
 }
